requirements.txt
requests==2.25.1
mailchimp-marketing==3.0.58
python-dotenv==0.15.0
